updated messenger page styling
