+++ client/src/components/convo/Convo.css
@@ -1,10 +1,9 @@
 /* Conversation Container */
 .convo {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	min-width: 40vh;
+	min-width: 70vw;
 	width: 100%;
+	/* min-width: 40vh;
+	width: 100%; */
 }
 
 /* Conversation Header */
@@ -14,15 +13,19 @@
 	align-items: center;
 	padding: 20px 10px;
 	background-color: white;
-	width: 100%;
-	height: 10vh;
+	filter: drop-shadow(0 0 5px #c1cbd9);
+
+	/* width: 100%;
+	height: 10vh; */
 }
 
 /* Conversation Body */
-.convoBody {
-	height: 78.5vh;
+.convoBodyContainer {
+	padding-left: 10px;
+}
+
+.convoContent {
 	padding: 0px 20px;
-	width: 100%;
 	overflow-y: auto;
 }
