+++ client/src/components/convo/children/ConvoHeader.js
@@ -7,9 +7,7 @@ const ConvoHeader = ({ currConvo }) => {
 	return (
 		<div className="convoHeader">
 			<DisplayAvatar id={currConvo._id} user={currConvo} width={6} height={6} />
-			<h2 style={{ marginLeft: "15px" }}>
-				{currConvo.firstName} {currConvo.lastName}
-			</h2>
+			<h2 style={{ marginLeft: "15px" }}>{currConvo.displayName}</h2>
 		</div>
 	);
 };
