+++ client/src/components/sidebar/children/Settings.js
@@ -2,12 +2,12 @@ import React, { useState } from "react";
 import { makeStyles } from "@material-ui/core/styles";
 import { useHistory } from "react-router-dom";
 import Modal from "@material-ui/core/Modal";
-import SettingsIcon from "@material-ui/icons/Settings";
+import MoreHorizIcon from "@material-ui/icons/MoreHoriz";
 import CancelIcon from "@material-ui/icons/Cancel";
 import Menu from "@material-ui/core/Menu";
 import MenuItem from "@material-ui/core/MenuItem";
 
-import ProfileForm from "../../ProfileForm";
+import EditDisplayName from "../modals/EditDisplayName";
 
 function getModalStyle() {
 	const top = 50;
@@ -58,10 +58,10 @@ const Settings = () => {
 	const body = (
 		<div style={modalStyle} className={classes.paper}>
 			<div className="modalHeader">
-				<h2>Edit Profile</h2>
+				<h2>Edit Display Name</h2>
 				<CancelIcon className="modalExit" onClick={() => setOpen(false)} />
 			</div>
-			<ProfileForm type="Update" route="/api/users/me" />
+			{/*	<EditDisplayName /> */}
 		</div>
 	);
 
@@ -72,10 +72,13 @@ const Settings = () => {
 				open={Boolean(anchorEl)}
 				onClose={() => setAnchorEl(null)}
 			>
-				<MenuItem onClick={() => setOpen(true)}>Edit Profile</MenuItem>
+				<MenuItem onClick={() => setOpen(true)}>Edit Display Name</MenuItem>
 				<MenuItem onClick={handleLogout}>Logout</MenuItem>
 			</Menu>
-			<SettingsIcon onClick={(e) => setAnchorEl(e.currentTarget)} />
+			<MoreHorizIcon
+				style={{ color: "#c9d1df" }}
+				onClick={(e) => setAnchorEl(e.currentTarget)}
+			/>
 			<Modal open={open}>{body}</Modal>
 		</div>
 	);
