+++ client/src/pages/Register.js
@@ -1,12 +1,21 @@
-import React, { useEffect } from "react";
+import React, { useState, useEffect } from "react";
 import { Link, useHistory } from "react-router-dom";
-import { Button, Box } from "@material-ui/core";
+import { Button, TextField, Box, Snackbar } from "@material-ui/core";
+import MuiAlert from "@material-ui/lab/Alert";
+import * as EmailValidator from "email-validator";
 
 import messengerIcon from "../assets/images/message.png";
-import ProfileForm from "../components/ProfileForm";
 import "./RegisterLogin.css";
 
+function Alert(props) {
+	return <MuiAlert elevation={6} variant="filled" {...props} />;
+}
+
 const Register = () => {
+	// State management for input and snackbars
+	const [inputValues, setInputValues] = useState({});
+	const [snackbar, setSnackbar] = useState({ open: false });
+
 	// Will be used to redirect user
 	const history = useHistory();
 
@@ -24,6 +33,112 @@ const Register = () => {
 		loginCheck();
 	}, [history]);
 
+	// Stores data in our input state
+	const handleChange = (e) => {
+		const { name, value } = e.target;
+		setInputValues({ ...inputValues, [name]: value });
+	};
+
+	// Sends data to our back end server
+	const handleSubmit = async () => {
+		// Checks for invalid email
+		if (!EmailValidator.validate(inputValues.email)) {
+			setSnackbar({
+				open: true,
+				severity: "error",
+				message: "Invalid email"
+			});
+			return;
+		}
+
+		// Password validator
+		if (inputValues.password) {
+			// Checks if they are at least 6 characters
+			if (inputValues.password.length < 6) {
+				setSnackbar({
+					open: true,
+					severity: "error",
+					message: "Password must be at least 6 characters"
+				});
+				return;
+			}
+		}
+
+		try {
+			// Our data to send to the server
+			const res = await fetch("api/register", {
+				method: "POST",
+				body: JSON.stringify(inputValues),
+				headers: { "Content-Type": "application/json" }
+			});
+
+			// Redirect them to messenger page
+			if (res.status === 201) {
+				history.push("/messenger");
+			}
+
+			// In the event we get an error
+			const data = await res.json();
+			if (data.errors) {
+				if (data.errors.email) {
+					setSnackbar({
+						open: true,
+						severity: "error",
+						message: data.errors.email
+					});
+				}
+			}
+		} catch (err) {
+			console.log(err);
+		}
+	};
+
+	// Pressing enter submits our form
+	const handleKeyPress = (e) => {
+		if (e.which === 13) {
+			handleSubmit();
+		}
+	};
+
+	// Closes our snackbar
+	const handleClose = (event, reason) => {
+		if (reason === "clickaway") {
+			return;
+		}
+
+		setSnackbar({ ...snackbar, open: false });
+	};
+
+	const snackbars = (
+		<div>
+			<Snackbar
+				anchorOrigin={{
+					vertical: "top",
+					horizontal: "center"
+				}}
+				open={snackbar.open}
+				autoHideDuration={6000}
+				onClose={handleClose}
+			>
+				<Alert onClose={handleClose} severity={snackbar.severity}>
+					{snackbar.message}
+				</Alert>
+			</Snackbar>
+		</div>
+	);
+
+	// Error handling for password (6 chars minimum) for material-ui
+	let validEmail, validPassword;
+
+	if (inputValues.email) {
+		validEmail = !EmailValidator.validate(inputValues.email);
+	}
+
+	if (inputValues.password || inputValues.confirmPassword) {
+		validPassword =
+			inputValues.password.length > 0 && inputValues.password.length < 6;
+	}
+
 	// JSX pertaining to our form
 	const formContainer = (
 		<div className="formContainer">
@@ -39,16 +154,61 @@ const Register = () => {
 					</Button>
 				</Link>
 			</Box>
-			<Box mt={5} mx={6}>
-				<h2>Create an account</h2>
-				<ProfileForm type="Register" route="api/register" />
+			<Box mt={10} mx={6}>
+				<h2 id="registerHeader">Create an account</h2>
+				<Box mt={2}>
+					<TextField
+						label="Display Name"
+						name="displayName"
+						onChange={(e) => handleChange(e)}
+						onKeyPress={handleKeyPress}
+						fullWidth
+					/>
+				</Box>
+				<Box mt={2}>
+					<TextField
+						label="Email"
+						name="email"
+						onChange={(e) => handleChange(e)}
+						onKeyPress={handleKeyPress}
+						error={validEmail}
+						helperText={validEmail ? "Enter a valid email address" : ""}
+						fullWidth
+					/>
+				</Box>
+				<Box mt={2} mb={4}>
+					<TextField
+						label="Password"
+						name="password"
+						type="password"
+						onChange={(e) => handleChange(e)}
+						onKeyPress={handleKeyPress}
+						error={validPassword}
+						helperText={
+							validPassword ? "Passwords must be at least 6 characters" : ""
+						}
+						fullWidth
+					/>
+				</Box>
+				<div className="authButton">
+					<Button
+						className="messengerButton"
+						onClick={handleSubmit}
+						size="large"
+						variant="contained"
+						color="primary"
+					>
+						Register
+					</Button>
+				</div>
+				{snackbars}
 			</Box>
 		</div>
 	);
 
 	return (
 		<div className="pageContainer">
-			<div className="sideContainer" style={{ marginBottom: "30px" }}>
+			<div className="sideContainer">
 				<div>
 					<img
 						id="messengerIcon"
